misplaced MultiIndex test
